# services/docs/openapi.yaml
openapi: "3.0.0"
info:
  title: AI-Carmine Broker API
  description: FastAPI application for agentic loop broker with job management, planner lab, and tool dispatch
  version: "1.0.0"
  contact:
    name: AI-Carmine
    url: https://github.com/C-F-tek/agentic-tool-loop
servers:
  - url: http://127.0.0.1:3572
    description: Vulkan Tool Broker (internal services)
paths:
...
  /planner-lab:
    get:
      summary: Planner lab index (HTML)
      tags: [planner-lab]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Planner lab index HTML

    post:
      summary: Start planner lab job
      tags: [planner-lab]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                task:
                  type: string
                request:
                  type: string
                prompt:
                  type: string
                return_mode:
                  type: string
                  enum: [background, wait, fire_and_forget]
                wait_seconds:
                  type: integer
                  minimum: 1
                  maximum: 30
                max_steps:
                  type: integer
                approval_mode:
                  type: string
                user_consent:
                  type: string
              required: [task]
      responses:
        "200":
          description: Agent response
          content:
            application/json:
              schema:
                type: object
...